+++ examples/b.js
@@ -2,3 +2,5 @@ function aaa( bbb ) {
    }
 function ccc   (   ) {
 ddd();}
+function eee(){fff();{ggg();{hhh();iii();}jjj();kkk();}}
+function ooo(){{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{{ppp();}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}