+++ examples/c.js
@@ -1,4 +1,7 @@
-const x=[1,2,3,4,5,6];
+const x=[1,2,3,4,5];
 const y={a:1,b:2,c:3,d:4};
 const z=[{a:1},{b:2},{c:3}];
-const xx=['aaaaaaaaaaaaaaaaaaaaaaa','bbbbbbbbbbbbbbbbbbbbbb','ccccccccccccccccccccccc','dddddddddddddddddddddddddddd'];
+const z=[{a:1,x:'aaaaaaaaaaaaaa'},{b:2,x:'aaaaaaaaaaaaaa'},{c:3,x:'aaaaaaaaaaaaaa'}];
+const a1=['aaaaaaaaaaaaaaaaaaaaaaa','bbbbbbbbbbbbbbbbbbbbbb','ccccccccccccccccccccccc','dddddddddddddddddddddddddddd'];
+const a2=[[1,2,3,4,5],[1,2,3,4,5],[1,2,3,4,5],[1,2,3,4,5],[1,2,3,4,5],[1,2,3,4,5],[1,2,3,4,5]];
+const o1={a:{x:100,y:101,z:102},b:{x:100,y:101,z:102},c:{x:100,y:101,z:102},d:{x:100,y:101,z:102},e:{x:100,y:101,z:102},}